feat : 예외처리 완료

+++ likelion-client/src/main/java/likelion/univ/project/dto/request/ProjectRequestDto.java
@@ -22,7 +22,7 @@ public class ProjectRequestDto {
     private String thon;
     private Output outPut;
     private String serviceName;
-    private int ordinal;
+    private long ordinal;
     private String univ;
     private LocalDate startDate;
     private LocalDate endDate;
